add global settings for refresh
